requirements/common.txt
Django==3.2.9
django-environ==0.8.1
Pillow==8.4.0
django-solo==2.0.0
django-tinymce==3.4.0
django-phonenumber-field==6.0.0
phonenumbers==8.12.40
psycopg2-binary==2.8.6
graphene-django==2.15.0
gunicorn==20.1.0
django-url-or-relative-url-field==0.2.0
django-admin-sortable2==1.0.3
graphene-file-upload==1.3.0
django-graphql-jwt==0.3.0
django-graphql-auth
# dependencies for neural network
numpy==1.22.2
protobuf==3.16.0
onnx==1.10.2
onnxruntime==1.10.0
albumentations==1.1.0
--find-links https://download.pytorch.org/whl/cpu/torch_stable.html
torch==1.10.2+cpu
--find-links https://download.pytorch.org/whl/cpu/torch_stable.html
torchvision==0.11.3+cpu
--find-links https://download.pytorch.org/whl/cpu/torch_stable.html
torchaudio==0.10.2+cpu